+++ src/site/markdown/CHANGES.md
@@ -3,6 +3,7 @@ CHANGES
 
 develop
 -------
+- show headers and cookies if the user decides it
 - Configure the encoding of the surefire JVM to UTF-8
 
 Version 0.6.1 19.10.2012